libndt: simpler-to-override resolver
Overriding getaddrinfo() seems a bit complex. We need to do that
in MK to use other resolvers. So, how about splitting connect_tcp()
such that name resolution fills a vector of addresses, using
getaddrinfo() by default, and then connect_tcp() iterates over
such vector using getaddrinfo() in a simpler way?

With this change, in MK we need to override a function that takes
in input a domain name (or IP address) and outputs the corresponding
list of IP addresses (or just the same IP). This is simpler to
implement with various DNS libraries than to write a full fledged
replacement for the whole complexity of getaddrinfo().