Implement netx layer that returns an error
Returning a clear error is better (in terms of code quality) than
encoding the error as a special case in the return value.

This would be of further help with SSL code, where we need to use the
return value to understand what kind of select() is required, i.e.
whether a SSL socket is blocked on reading or writing.