libndt: make select() more secure
There is the issue that in principle select() may cause segfaults
if one sets more than FD_SETSIZE file descriptors.

The code checking that we were correctly using select() did not
perform these checks and was messy.

So, refactor this code into a simpler to use interface and ensure
that all FD_SETSIZE checks are performed in the right place.