poll: more robust flags handling (#56)
After a couple of assertions triggered, I realized that poll() does
not necessarily set POLLIN on EOF or error if you requested for
POLLIN (and similarly for POLLOUT). So, the most robust behavior
is that of actually just checking whether there are flags and then
attempting to perform some I/O to see a result or an error.