libndt-client: stop ignoring SIGPIPE
To make sure we are correctly ignoring SIGPIPE in a thread
specific way, let's start with not ignoring SIGPIPE on a
per process basis in the test client.

This will hopefully cause some Travis tests to fail. Then we
can add support for ignoring SIGPIPE for just the thread that
is running NDT using a system specific mechanism.