Just ignore SIGPIPE in the thread running a test [in progress] (#58)
* libndt-client: stop ignoring SIGPIPE

To make sure we are correctly ignoring SIGPIPE in a thread
specific way, let's start with not ignoring SIGPIPE on a
per process basis in the test client.

This will hopefully cause some Travis tests to fail. Then we
can add support for ignoring SIGPIPE for just the thread that
is running NDT using a system specific mechanism.

* libndt: ignore SIGPIPE on BSD systems

* Use OpenSSL bio to enforce MSG_NOSIGNAL on Linux

Adapted from the way in which librabbitmq handles the same
issue. See the code comments for more details.

While preparing this diff, I also got confirmation of a bug
during the upload phase that I did not except to see.

Such bug has been documented in a code comment and is something
I will investigate once the current PR has landed.

* Add missing include; repairs GCC and MSVC builds

* fix windows build for real

* [ci skip] remove file I did not mean to commit

* [ci skip] fully comply with MIT license

* [ci skip] Tweak comment

* [ci skip] further tweak comment for more clarity

* [ci skip] more precise comment

* Also include atomic

* Make sure we'll not cause SIGPIPE