Enforce constness in s2c and c2s threads (#72)
Basically, make sure that threads have constant access only to the class variables. This means making sure that applying the const attribute to the class also propagates to the class variables (i.e. no pimpl), passing a const pointer to the threads lambda, and properly tagging as const all the methods that are used inside the lambda. All of this gives us the guarantee that no modification can be made. Plus there should be no race since values are not changed when threads are alive.