ndt7: resolve build issue with GCC
I already run into this issue some time ago. With GCC what
basically happens is that `~` and `&` operate on integers and
not on chars. So, using that operation with chars is going
to create a warning that for us is an error due to `-Werror`.

This could be solved using casting. However, there is little
point in introducing obscure code. It would equally be OK
to issue a warning if an unrecognized flag is set.