ndt7: report the integral of the speed
It used to be different for ndt5, where we reported the speed
measured in each bucket. The reason why ndt5 is different is
that we wanted the speed in each interval to sort out all the
speeds and exclude the extreme samples. This was done by MK
to report to users a speed that was closer to the one that it
would be reported by OOKLA's SpeedTest without having to use
several parallel network connections. Now that we have BBR
in the backend for ndt7, instead, we can use the integral of
the speed, which is less prone to fluctuations.