ndt7: implement the upload subtest (#87)
* ndt7: sketch out uploader

* ndt7: move #ifdef LIBNDT_HAVE_OPENSSL check above

Allows to have a ndt7 implementation that is not riddled
with ifdef and endif statements.

* ndt7: factor common code for connecting

* ndt7: measure download and upload goodput

* ndt7: tweak what we print by default

* ndt7: prepare frame to speed up upload

* ndt7: keep going in case of errors

* ndt7: report the integral of the speed

It used to be different for ndt5, where we reported the speed
measured in each bucket. The reason why ndt5 is different is
that we wanted the speed in each interval to sort out all the
speeds and exclude the extreme samples. This was done by MK
to report to users a speed that was closer to the one that it
would be reported by OOKLA's SpeedTest without having to use
several parallel network connections. Now that we have BBR
in the backend for ndt7, instead, we can use the integral of
the speed, which is less prone to fluctuations.

* ndt7: explicitly print BBR measurements

I am not quite satisfied with the output yet. So, I will try
to make more changes to better understand what I see.

* ndt7: optionally send/recv feedback messages

* ndt7: implement the basic client flow

* ndt5: use prepared frames for legacy tests

* We don't need a ndt7 specific callback

* Better name for function

* Better comment

* Uniform handling of the performance callback

Just always report cumulative numbers. A ndt5 client that wants
to have per measurement interval stats can perform the derivative
for itself. This removes the ambiguity with respect to the
meaning of the results between ndt5 and ndt7.