Convert libndt to use mkbuild
This diff converts libndt to use mkbuild as opposed to deprecated
subrepositories that we would like to decommission.

It includes uncommitted changes to mkbuild required to make this conversion
actually possible, see https://github.com/measurement-kit/mkbuild/pull/6.

It also removes conditional support for OpenSSL and cURL mainly
because that simplifies the migration, _and_ because having such
support as optional was optional complexity.

This diff additionally changes the testing strategy in Travis
to only cover the legacy code path. It appears that there have
been changes again and now running several NDT tests from the
same invocation of Travis has become problematic.

This is reasonably okay because at this stage we're going to use
libndt only as an implementation of the legacy protocol.

Closes #91.

Closes #92.