Major refactoring of libndt
- structure the codebase like nlohmann/json

- move internal headers in include/libndt/internal

- move public headers in include/libndt

- ditch the option to disable compiling inline (too much pain)

- vendor all the dependencies we can vendor in third_party

- assemble the single include in single_include and add to it
also nlohmann/json so it basically becomes standalone

- get rid of MKBuild and curate the last generated CMakeLists.txt
such that it makes sense

The objective of this refactoring is to enable easier development
with clearly defined interfaces. We're not quite there but we
are getting closer and the structure we have now should be easier
to understand and more clear than the one we had before.